refactor(storage-test-utils): move FakeStorageService out of common
