+++ libs/storage-test-utils/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./fake-storage.service";
